+++ README.md
@@ -1,1 +1,115 @@
-# safehouse
+# SafeHouse🏠
+## grassroots initiative for earthquake safety
+
+## Check out a demo of the site here: 
+
+# Nathan Chan | July 29-30, 2019
+## General Assembly - Two Day Hackathon
+### Placed third overall
+### Worked with two UX/UI designers and two other software developers
+### The app was developed using the React.js framework. 
+
+# Introduction
+Given the most recent earthquake in Monroe, WA and the increasing likelihood of earthquakes occuring in the Greater Seattle Area, our team created an app that would allow users to identified if their Unreinforced Masonry Building (URM) is at medium risk, high risk, or critical risk in an earthquake. The project pulled from the City of Seattle's Public URM API. 
+
+# Planning
+Over the course of two days we were tasked with creating an app which uses a public API and relates to the topic of housing in Seattle. 
+
+Our team spent the first hour generating a proposal and pitching the idea to our cohort mentors. We finally landed on the idea of SafeHouse, a grassroots initiative for earthquake safety. 
+
+Since our team had a couple of designers, they took the lead in developing the wireframes, design, branding, and presentation. The developers turned those sketches and ideas and turned them into a reality. 
+
+## URM API
+
+https://data.seattle.gov/Permitting/Unreinforced-Masonry-Buildings/54qs-2h7f
+
+
+## UX/UI 
+
+
+## React.js
+
+
+# Live Demo and Presentation
+
+
+# Next Steps
+
+
+
+# Credit To 
+
+
+
+
+
+
+
+This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+
+## Available Scripts
+
+In the project directory, you can run:
+
+### `npm start`
+
+Runs the app in the development mode.<br>
+Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
+
+The page will reload if you make edits.<br>
+You will also see any lint errors in the console.
+
+### `npm test`
+
+Launches the test runner in the interactive watch mode.<br>
+See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
+
+### `npm run build`
+
+Builds the app for production to the `build` folder.<br>
+It correctly bundles React in production mode and optimizes the build for the best performance.
+
+The build is minified and the filenames include the hashes.<br>
+Your app is ready to be deployed!
+
+See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+
+### `npm run eject`
+
+**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
+
+If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
+
+Instead, it will copy all the configuration files and the transitive dependencies (Webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
+
+You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
+
+## Learn More
+
+You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
+
+To learn React, check out the [React documentation](https://reactjs.org/).
+
+### Code Splitting
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/code-splitting
+
+### Analyzing the Bundle Size
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size
+
+### Making a Progressive Web App
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app
+
+### Advanced Configuration
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/advanced-configuration
+
+### Deployment
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/deployment
+
+### `npm run build` fails to minify
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify
